move files to better places
