+++ packages/react-pages/src/node/page-strategy/UpdateBuffer.ts
