+++ packages/react-pages/src/node/virtule-module-plugins/ts-info-module/index.ts
@@ -41,3 +41,5 @@ export class TsInfoModuleManager {
     this.pmm.close()
   }
 }
+
+export * from './mdx-plugin'
